requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64

django==3.2.8
djangorestframework==3.12.4